+++ src/languages/QOwnNotes_es.ts
@@ -4531,7 +4531,7 @@ Se debe probar y si hay conflictos de sincronización establecer un valor más a
     <message>
       <location filename="../dialogs/settingsdialog.ui" line="2540"/>
       <source>Enforce system icon theme</source>
-      <translation type="unfinished">Enforce system icon theme</translation>
+      <translation>Forzar el tema de iconos de sistema</translation>
     </message>
     <message>
       <location filename="../dialogs/settingsdialog.ui" line="2636"/>