+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # QOwnNotes Changelog
 
+## 19.4.5
+- more checksum generation improvements for the AUR and added checksum support
+  to the Launchpad Snap build process (for [#1175](https://github.com/pbek/QOwnNotes/issues/1175))
+
 ## 19.4.4
 - fixed renaming issue on case-insensitive file-systems
   (for [#1190](https://github.com/pbek/QOwnNotes/issues/1190))